+++ src/components/branchlist.rs
@@ -7,17 +7,18 @@ use crate::{
     keys::SharedKeyConfig,
     queue::{Action, InternalEvent, NeedsUpdate, Queue},
     strings, try_or_popup,
-    ui::{self, calc_scroll_top},
+    ui::{self, calc_scroll_top, Size},
 };
+use anyhow::Result;
 use asyncgit::{
-    sync::{checkout_branch, get_branches_info, BranchInfo},
+    sync::{
+        branch::checkout_remote_branch, checkout_branch,
+        get_branches_info, BranchInfo,
+    },
     CWD,
 };
 use crossterm::event::Event;
-use std::{
-    cell::Cell,
-    convert::{TryFrom, TryInto},
-};
+use std::{cell::Cell, convert::TryInto};
 use tui::{
     backend::Backend,
     layout::{Alignment, Rect},
@@ -25,15 +26,13 @@ use tui::{
     widgets::{Block, BorderType, Borders, Clear, Paragraph},
     Frame,
 };
-use unicode_truncate::UnicodeTruncateStr;
-
-use crate::ui::Size;
-use anyhow::Result;
 use ui::style::SharedTheme;
+use unicode_truncate::UnicodeTruncateStr;
 
 ///
 pub struct BranchListComponent {
-    branch_names: Vec<BranchInfo>,
+    branches: Vec<BranchInfo>,
+    local: bool,
     visible: bool,
     selection: u16,
     scroll_top: Cell<usize>,
@@ -80,7 +79,7 @@ impl DrawableComponent for BranchListComponent {
                 ))
                 .block(
                     Block::default()
-                        .title(strings::SELECT_BRANCH_POPUP_MSG)
+                        .title(strings::title_branches(self.local))
                         .border_type(BorderType::Thick)
                         .borders(Borders::ALL),
                 )
@@ -92,11 +91,11 @@ impl DrawableComponent for BranchListComponent {
                 f,
                 area,
                 &self.theme,
-                self.branch_names.len(),
+                self.branches.len(),
                 self.scroll_top.get(),
             );
 
-            self.current_height.set(u16::try_from(height_in_lines)?);
+            self.current_height.set(height_in_lines.try_into()?);
         }
 
         Ok(())
@@ -125,15 +124,7 @@ impl Component for BranchListComponent {
             ));
 
             out.push(CommandInfo::new(
-                strings::commands::open_branch_create_popup(
-                    &self.key_config,
-                ),
-                true,
-                true,
-            ));
-
-            out.push(CommandInfo::new(
-                strings::commands::delete_branch_popup(
+                strings::commands::select_branch_popup(
                     &self.key_config,
                 ),
                 !self.selection_is_cur_branch(),
@@ -141,8 +132,33 @@ impl Component for BranchListComponent {
             ));
 
             out.push(CommandInfo::new(
+                strings::commands::open_branch_create_popup(
+                    &self.key_config,
+                ),
+                true,
+                self.local,
+            ));
+
+            out.push(CommandInfo::new(
+                strings::commands::delete_branch_popup(
+                    &self.key_config,
+                ),
+                !self.selection_is_cur_branch(),
+                self.local,
+            ));
+
+            out.push(CommandInfo::new(
                 strings::commands::rename_branch_popup(
                     &self.key_config,
+                ),
+                true,
+                self.local,
+            ));
+
+            out.push(CommandInfo::new(
+                strings::commands::toggle_branch_popup(
+                    &self.key_config,
+                    self.local,
                 ),
                 true,
                 true,
@@ -170,8 +186,6 @@ impl Component for BranchListComponent {
                         "switch branch error:",
                         self.switch_to_selected_branch()
                     );
-
-                    self.hide()
                 } else if e == self.key_config.create_branch {
                     self.queue
                         .borrow_mut()
@@ -179,7 +193,7 @@ impl Component for BranchListComponent {
                     self.hide();
                 } else if e == self.key_config.rename_branch {
                     let cur_branch =
-                        &self.branch_names[self.selection as usize];
+                        &self.branches[self.selection as usize];
                     self.queue.borrow_mut().push_back(
                         InternalEvent::RenameBranch(
                             cur_branch.reference.clone(),
@@ -194,13 +208,17 @@ impl Component for BranchListComponent {
                     self.queue.borrow_mut().push_back(
                         InternalEvent::ConfirmAction(
                             Action::DeleteBranch(
-                                self.branch_names
+                                self.branches
                                     [self.selection as usize]
                                     .reference
                                     .clone(),
                             ),
                         ),
                     );
+                } else if e == self.key_config.toggle_remote_branches
+                {
+                    self.local = !self.local;
+                    self.update_branches()?;
                 }
             }
 
@@ -232,7 +250,8 @@ impl BranchListComponent {
         key_config: SharedKeyConfig,
     ) -> Self {
         Self {
-            branch_names: Vec::new(),
+            branches: Vec::new(),
+            local: true,
             visible: false,
             selection: 0,
             scroll_top: Cell::new(0),
@@ -253,17 +272,22 @@ impl BranchListComponent {
 
     /// fetch list of branches
     pub fn update_branches(&mut self) -> Result<()> {
-        self.branch_names = get_branches_info(CWD)?;
+        self.branches = get_branches_info(CWD, self.local)?;
         self.set_selection(self.selection)?;
         Ok(())
     }
 
     fn selection_is_cur_branch(&self) -> bool {
-        self.branch_names
+        self.branches
             .iter()
             .enumerate()
             .filter(|(index, b)| {
-                b.is_head && *index == self.selection as usize
+                b.local_details()
+                    .map(|details| {
+                        details.is_head
+                            && *index == self.selection as usize
+                    })
+                    .unwrap_or_default()
             })
             .count()
             > 0
@@ -289,7 +313,7 @@ impl BranchListComponent {
     }
 
     fn set_selection(&mut self, selection: u16) -> Result<()> {
-        let num_branches: u16 = self.branch_names.len().try_into()?;
+        let num_branches: u16 = self.branches.len().try_into()?;
         let num_branches = num_branches.saturating_sub(1);
 
         let selection = if selection > num_branches {
@@ -326,7 +350,7 @@ impl BranchListComponent {
         let mut txt = Vec::new();
 
         for (i, displaybranch) in self
-            .branch_names
+            .branches
             .iter()
             .skip(self.scroll_top.get())
             .take(height)
@@ -354,9 +378,16 @@ impl BranchListComponent {
             let selected =
                 self.selection as usize - self.scroll_top.get() == i;
 
-            let is_head_str =
-                if displaybranch.is_head { "*" } else { " " };
-            let has_upstream_str = if displaybranch.has_upstream {
+            let is_head = displaybranch
+                .local_details()
+                .map(|details| details.is_head)
+                .unwrap_or_default();
+            let is_head_str = if is_head { "*" } else { " " };
+            let has_upstream_str = if displaybranch
+                .local_details()
+                .map(|details| details.has_upstream)
+                .unwrap_or_default()
+            {
                 "\u{2191}"
             } else {
                 " "
@@ -383,7 +414,7 @@ impl BranchListComponent {
                     branch_name,
                     w = branch_name_length
                 ),
-                theme.branch(selected, displaybranch.is_head),
+                theme.branch(selected, is_head),
             );
 
             txt.push(Spans::from(vec![
@@ -398,11 +429,22 @@ impl BranchListComponent {
     }
 
     ///
-    fn switch_to_selected_branch(&self) -> Result<()> {
-        checkout_branch(
-            asyncgit::CWD,
-            &self.branch_names[self.selection as usize].reference,
-        )?;
+    fn switch_to_selected_branch(&mut self) -> Result<()> {
+        if self.local {
+            checkout_branch(
+                asyncgit::CWD,
+                &self.branches[self.selection as usize].reference,
+            )?;
+            self.hide()
+        } else {
+            checkout_remote_branch(
+                CWD,
+                &self.branches[self.selection as usize],
+            )?;
+            self.local = true;
+            self.update_branches()?;
+        }
+
         self.queue
             .borrow_mut()
             .push_back(InternalEvent::Update(NeedsUpdate::ALL));